// packages/cwsandbox/src/transports/node-grpc/generated/coreweave/sandbox/v1beta2/secrets.ts
export interface SecretStoreReference {
    /**
     * Name of a configured secret store (resolved server-side).
     *
     * @generated from protobuf field: string store_name = 1
     */
    storeName: string;
    /**
     * Secrets to fetch from this store and inject into the sandbox.
     *
     * @generated from protobuf field: repeated coreweave.sandbox.v1beta2.SecretMapping secrets = 2
     */
    secrets: SecretMapping[];
}
/**
 * SecretMapping maps a provider-specific secret path to an environment variable.
 *
 * @generated from protobuf message coreweave.sandbox.v1beta2.SecretMapping
 */
export interface SecretMapping {
    /**
     * Provider-specific secret identifier (e.g., "HF_TOKEN", "secret/data/myapp/db").
     *
     * @generated from protobuf field: string path = 1
     */
    path: string;
    /**
     * Optional field within a structured secret (for key-value secrets).
     *
     * @generated from protobuf field: string field = 2
     */
    field: string;
    /**
     * Environment variable name for delivery into the sandbox container.
     *
     * @generated from protobuf field: string env_var = 3
     */
    envVar: string;
}
// ...
class SecretStoreReference$Type extends MessageType<SecretStoreReference> {
    constructor() {
        super("coreweave.sandbox.v1beta2.SecretStoreReference", [
            { no: 1, name: "store_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "secrets", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => SecretMapping }
        ]);
    }
    create(value?: PartialMessage<SecretStoreReference>): SecretStoreReference {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.storeName = "";
        message.secrets = [];
        if (value !== undefined)
            reflectionMergePartial<SecretStoreReference>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SecretStoreReference): SecretStoreReference {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string store_name */ 1:
                    message.storeName = reader.string();
                    break;
                case /* repeated coreweave.sandbox.v1beta2.SecretMapping secrets */ 2:
                    message.secrets.push(SecretMapping.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SecretStoreReference, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string store_name = 1; */
        if (message.storeName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.storeName);
        /* repeated coreweave.sandbox.v1beta2.SecretMapping secrets = 2; */
        for (let i = 0; i < message.secrets.length; i++)
            SecretMapping.internalBinaryWrite(message.secrets[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message coreweave.sandbox.v1beta2.SecretStoreReference
 */
export const SecretStoreReference = new SecretStoreReference$Type();
// @generated message type with reflection information, may provide speed optimized methods
class SecretMapping$Type extends MessageType<SecretMapping> {
    constructor() {
        super("coreweave.sandbox.v1beta2.SecretMapping", [
            { no: 1, name: "path", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "field", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "env_var", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<SecretMapping>): SecretMapping {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.path = "";
        message.field = "";
        message.envVar = "";
        if (value !== undefined)
            reflectionMergePartial<SecretMapping>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SecretMapping): SecretMapping {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string path */ 1:
                    message.path = reader.string();
                    break;
                case /* string field */ 2:
                    message.field = reader.string();
                    break;
                case /* string env_var */ 3:
                    message.envVar = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SecretMapping, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string path = 1; */
        if (message.path !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.path);
        /* string field = 2; */
        if (message.field !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.field);
        /* string env_var = 3; */
        if (message.envVar !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.envVar);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message coreweave.sandbox.v1beta2.SecretMapping
 */
export const SecretMapping = new SecretMapping$Type();